Moving contact view tests from eyb into international
