+++ pii-secret-exclude.txt
@@ -149,3 +149,4 @@ international/migrations/0001_initial.py
 tests/unit/sso_profile/business_profile/test_helpers.py
 tests/unit/core/management/commands/test_send_campaign_site_review_reminder.py
 tests/unit/international/test_forms.py
+tests/unit/international/test_views.py
